Added validation

+++ app/routes/category/category.py
@@ -1,6 +1,7 @@
 from app import app, mongo
 from app.utils.objectIdToStr import objectIdToStr
 from app.models.Category import Category
+from app.models.Part import Part
 from bson import ObjectId
 from flask import request, jsonify
 
@@ -9,7 +10,7 @@ from flask import request, jsonify
 def add_category():
     try:
         data = request.json
-        Category.validate(data)
+        Category.validate(data, mongo)
         new_category = Category(**data)
         mongo.db.categories.insert_one(new_category.to_dict())
         return jsonify({'msg': 'Category added successfully'}), 201
@@ -37,11 +38,14 @@ def update_category(id):
         return jsonify({'error': 'Invalid ObjectId'}), 400
 
     data = request.json
+    if '_id' in data and data['_id'] != id:
+        return jsonify({'error': 'Mismatch between URL id and JSON id'}), 400
+    
     try:
-        Category.validate(data)
+        Category.validate(data, mongo)
     except ValueError as e:
         return jsonify({'error': str(e)}), 400
-
+    
     result = mongo.db.categories.update_one({'_id': object_id}, {"$set": data})
     if result.matched_count:
         return jsonify({'msg': 'Category updated successfully'}), 200
@@ -55,6 +59,12 @@ def delete_category(id):
         object_id = ObjectId(id)
     except:
         return jsonify({'error': 'Invalid ObjectId'}), 400
+
+    if Part.category_has_parts(id, mongo):
+        return jsonify({'error': 'Cannot delete category as it has parts assigned to it'}), 400
+    
+    if Part.has_parts_in_subtree(id, mongo):
+        return jsonify({'error': 'Cannot delete category because some of their children have parts assigned to them'}), 400
 
     result = mongo.db.categories.delete_one({'_id': object_id})
     if result.deleted_count:
